+++ mrblib/itamae.rb
@@ -1,3 +1,0 @@
-def __main__(argv)
-  Itamae::CLI.start(argv)
-end