+++ mrblib/mitamae.rb
@@ -1,0 +1,3 @@
+def __main__(argv)
+  MItamae::CLI.start(argv)
+end