+++ mrblib/mitamae/node.rb
@@ -1,0 +1,60 @@
+module MItamae
+  class Node
+    ValidationError = Class.new(StandardError)
+
+    attr_reader :mash
+
+    def initialize(hash, backend)
+      @mash = Hashie::Mash.new(hash)
+      @backend = backend
+    end
+
+    def reverse_merge(other_hash)
+      self.class.new(_reverse_merge(other_hash), @backend)
+    end
+
+    def reverse_merge!(other_hash)
+      @mash.replace(_reverse_merge(other_hash))
+    end
+
+    def [](key)
+      if @mash.has_key?(key)
+        @mash[key]
+      else
+        fetch_inventory_value(key)
+      end
+    end
+
+    private
+
+    def _reverse_merge(other_hash)
+      Hashie::Mash.new(other_hash).merge(@mash)
+    end
+
+    def method_missing(method, *args)
+      if @mash.respond_to?(method)
+        return @mash.public_send(method, *args)
+      elsif args.empty? && value = fetch_inventory_value(method)
+        return value
+      end
+
+      super
+    end
+
+    def respond_to?(method, priv = false)
+      @mash.respond_to?(method, priv) || super
+    end
+
+    def fetch_inventory_value(key)
+      host_inventory[key]
+    end
+
+    def host_inventory
+      @host_inventory ||= Hashie::Mash.new.tap do |inv|
+        os_info = @backend.os_info
+        inv[:platform] = os_info[:family]
+        inv[:platform_version] = os_info[:release]
+      end
+    end
+  end
+end