+++ mrblib/mitamae/resource/file.rb
@@ -1,0 +1,15 @@
+module MItamae
+  module Resource
+    class File < Base
+      define_attribute :action, default: :create
+      define_attribute :path, type: String, default_name: true
+      define_attribute :content, type: String, default: nil
+      define_attribute :mode, type: String
+      define_attribute :owner, type: String
+      define_attribute :group, type: String
+      define_attribute :block, type: Proc, default: proc {}
+
+      self.available_actions = [:create, :delete]
+    end
+  end
+end