+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## v0.3.0
-- Support resource plugin [#4](https://github.com/k0kubun/itamae-mruby/pull/4)
+- Support resource plugin [#4](https://github.com/k0kubun/mitamae/pull/4)
 
 ## v0.2.2
 
@@ -50,8 +50,8 @@
 
 ## v0.0.10
 
-- Add group resource [#2](https://github.com/k0kubun/itamae-mruby/pull/2). *Thanks to @nonylene*
-- Add user resource [#3](https://github.com/k0kubun/itamae-mruby/pull/3). *Thanks to @nonylene*
+- Add group resource [#2](https://github.com/k0kubun/mitamae/pull/2). *Thanks to @nonylene*
+- Add user resource [#3](https://github.com/k0kubun/mitamae/pull/3). *Thanks to @nonylene*
 
 ## v0.0.9